python-fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2
python-dotenv
unidecode